Refactored code, Added some more components.
