+++ .vscode/settings.json
@@ -9,6 +9,9 @@
     // Turn off tsc task auto detection since we have the necessary tasks as npm scripts
     "typescript.tsc.autoDetect": "off",
     "cSpell.words": [
+        "flexipage",
+        "modulesexp",
+        "openinorg",
         "openiteminorg",
         "refreshinstanceurl"
     ]
